iteration -4 complete
